✨ 리뷰 반영 - useAuthValidation 삭제 - invalidate account api 추가 - 카테고리/collection api 원래 자리로 - 스켈레톤 밈 디자인 수정 [QA-2]
