+++ src/application/hooks/api/collection/index.ts
@@ -1,10 +1,13 @@
-import type { QueryFunctionContext, UseQueryOptions } from "@tanstack/react-query";
-import { useInfiniteQuery, useMutation, useQuery, useQueryClient } from "@tanstack/react-query";
+import type { UseQueryOptions } from "@tanstack/react-query";
+import { useMutation, useQuery, useQueryClient } from "@tanstack/react-query";
 
 import { api } from "@/infra/api";
 import type { GetCollectionCheckResponse } from "@/infra/api/collection/types";
 
+import { QUERY_KEYS as ACCOUNT_QUERY_KEYS } from "../account/queryKey";
+import { QUERY_KEYS as MEME_QUERY_KEYS } from "../meme/queryKey";
 import { QUERY_KEYS } from "./queryKey";
+
 /**
  * 밈별 콜렉션 정보 API
  */
@@ -86,32 +89,6 @@ export const usePostMemeToCollection = () => {
   });
 };
 
-/**
- * 콜렉션 별 밈 리스트 API
- *
- * NOTE
- * 현재 하나의 콜렉션만이 존재(즉, collectionId가 하나)
- * 추후에 여러개의 콜렉션을 다룰 예정(즉, collectionId에 여러개)
- *
- * 마이페이지(/mypage)에서는 무한스크롤 적용 안함
- * 콜렉션 페이지(/collect) 페이지에서 무한 스크롤 적용함
- */
-export const useGetMemesByCollectionId = (collectionId: number) => {
-  const { data, fetchNextPage } = useInfiniteQuery({
-    queryKey: QUERY_KEYS.getMemesByCollectionId(collectionId),
-    queryFn: ({ pageParam = 0 }: QueryFunctionContext) =>
-      api.meme.getMemesByCollectionId({ collectionId, offset: pageParam, limit: 10 }),
-    getNextPageParam: (lastPage) => {
-      const { isLastPage, offset, limit } = lastPage;
-      return isLastPage ? undefined : offset + limit;
-    },
-  });
-  const memeList = data ? data.pages.flatMap(({ data }) => data) : [];
-  const isEmpty = data?.pages[0].data.length === 0;
-
-  return { data: memeList, fetchNextPage, isEmpty };
-};
-
 export const usePostMemeToSharedCollection = ({
   memeId,
   sharedId,
@@ -124,7 +101,8 @@ export const usePostMemeToSharedCollection = ({
   return useMutation({
     mutationFn: () => api.collection.postMemeToSharedCollection(memeId),
     onSuccess: () => {
-      queryClient.invalidateQueries({ queryKey: QUERY_KEYS.getMemesByCollectionId(sharedId) });
+      queryClient.invalidateQueries({ queryKey: MEME_QUERY_KEYS.getMemesByCollectionId(sharedId) });
+      queryClient.invalidateQueries({ queryKey: ACCOUNT_QUERY_KEYS.getMyAccount });
     },
   });
 };
