+++ src/application/hooks/api/collection/queryKey.ts
@@ -1,4 +1,3 @@
 export const QUERY_KEYS = {
   getCollectionCheck: (memeId: number) => ["getCollectionCheck", memeId],
-  getMemesByCollectionId: (collectionId: number | null) => ["getMemesByCollectionId", collectionId],
 } as const;
