+++ src/application/hooks/domain/auth/index.ts
@@ -1,2 +1,1 @@
 export * from "./useAuth";
-export * from "./useAuthValidation";
